+++ libs/ngx-material-intl-tel-input-lib/src/index.ts
@@ -3,3 +3,4 @@ export * from './lib/components/phone-icon/phone-icon.component';
 export * from './lib/types/text-labels.type';
 export * from './lib/enums/country-iso.enum';
 export * from './lib/services/country-display-name/country-display-name.service';
+export * from './lib/validators/tel-signal.validators';